+++ CHANGELOG.md
@@ -9,6 +9,10 @@
 - Elements now provide a `findRecursive` method allowing you to recursively
   find matching elements.
 
+### Array Element
+
+- New `isEmpty` convenience property for determining if an array is empty.
+
 # 0.15.0 - 2017-04-03
 
 - Getters of link element will now return an element